refactor(db): move each DB handler in its own file
