+++ reputation/api/tests/test_spamhaus_service.py
@@ -17,7 +17,7 @@ class TestSpamhausService(unittest.TestCase):
         now = datetime.now()
         last_year = now.replace(year=now.year - 1)
 
-        with patch('db.db.Postgres') as mock:
+        with patch('db.postgres.Postgres') as mock:
             instance = mock.return_value
             enter = instance.__enter__.return_value
             enter.find_spamhaus_entries.return_value = [{
